requirements.txt
argon2-cffi==21.1.0
asgiref==3.2.10
astroid==2.6.6
bcrypt==3.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-bootstrap3==21.1
django-braces==1.15.0
django-ckeditor==6.2.0
django-classy-tags==3.0.1
django-cms==3.9.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.2
django-decouple==2.1
django-environ==0.8.1
django-filter==21.1
django-formtools==2.3
django-heroku==0.3.1
django-js-asset==1.2.2
django-rest-swagger==2.2.0
django-sekizai==3.0.1
django-taggit==1.2.0
django-treebeard==4.5.1
django-utils-six==2.0
django-widget-tweaks==1.4.12
djangocms-admin-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
drf-yasg==1.20.0
factory-boy==2.12.0
Faker==4.1.0
gunicorn==20.0.4
html5lib==1.1
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.6.0
Markdown==3.3.6
MarkupSafe==2.0.1
mccabe==0.6.1
misaka==2.1.1
openapi-codec==1.3.2
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.21
pyflakes==2.2.0
pylint==2.9.6
pyparsing==3.0.7
python-dateutil==2.8.1
python-decouple==3.5
pytz==2020.1
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.14.0
soupsieve==2.3.1
sqlparse==0.3.1
static3==0.7.0
text-unidecode==1.3
toml==0.10.2
typing-extensions
uritemplate==4.1.1
urllib3==1.26.8
webencodings==0.5.1
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.7.0
